prevent bad compression that produces files larger than the original
For this I simply check after compression if the compressed data was
actually larger (or the same size) than the original uncompressed file.
If it was I drop the compressed data and write an uncompressed file
instead.

This might happen for files that are compressed alread. Compressing
those again will add overhead and as such would result in larger files
than the original.